bug fixes for logging

+++ Logger.h
@@ -185,7 +185,7 @@ template< typename T, typename FLT >
 inline void Logger<T,FLT>::op2( uint16_t op, const T * opnd1, const T& opnd2 )
 {
     if ( out_text ) {
-        *out << "op2i( " << op_to_str( op ) << ", " << opnd1 << ", 0x" << std::hex << opnd2 << " )\n";
+        *out << "op2i( " << op_to_str( op ) << ", " << opnd1 << ", 0x" << std::hex << int64_t(opnd2) << " )\n";
     }
 }
